remove convention in favour of selecting classes in the register
